+++ CHANGELOG.md
@@ -5,22 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- plugin state API
+
 ## [0.4.0]
 
 ### Added
 - Added the ability to change the text color, background and font thickness in the table using metadata
 
-## Changed
+### Changed
 - Refactoring for new SDK version
 
 ## [0.3.0]
 
-## Added
+### Added
 
 - logging
 - missing data message
 
-## Changed
+### Changed
 
 - layout styles
 - panel config logic